web/requirements.txt
fastapi==0.112.0
uvicorn==0.30.6
jinja2==3.1.4
ldap3==2.9.1
python-multipart==0.0.9
bcrypt==4.2.0
passlib==1.7.4
motor==3.5.1
pyyaml==6.0.2
colorlog==6.8.2
beanie==1.26.0
pyjwt==2.9.0